Extract UpdateObjectService from CocinaObjectStore
Now CocinaObjectStore has fewer dependencies (e.g. EventFactory) and is only concerned with operations on the datastore.
