+++ .rubocop_todo.yml
@@ -1,6 +1,6 @@
 # This configuration was generated by
 # `rubocop --auto-gen-config`
-# on 2022-05-10 17:43:38 UTC using RuboCop version 1.29.0.
+# on 2022-06-10 21:06:23 UTC using RuboCop version 1.30.1.
 # The point is for the user to remove these configuration records
 # one by one as the offenses are removed from the code base.
 # Note that changes in the inspected code, or installation of new
@@ -12,7 +12,7 @@ Lint/SuppressedException:
   Exclude:
     - 'lib/tasks/rspec.rake'
 
-# Offense count: 58
+# Offense count: 57
 # Configuration parameters: IgnoredMethods, CountRepeatedAttributes.
 Metrics/AbcSize:
   Max: 39
@@ -22,7 +22,7 @@ Metrics/AbcSize:
 Metrics/ClassLength:
   Max: 114
 
-# Offense count: 10
+# Offense count: 11
 # Configuration parameters: IgnoredMethods.
 Metrics/CyclomaticComplexity:
   Max: 11
@@ -30,7 +30,7 @@ Metrics/CyclomaticComplexity:
 # Offense count: 56
 # Configuration parameters: CountComments, CountAsOne, ExcludedMethods, IgnoredMethods.
 Metrics/MethodLength:
-  Max: 43
+  Max: 28
 
 # Offense count: 2
 # Configuration parameters: CountKeywordArgs, MaxOptionalParameters.
@@ -40,7 +40,7 @@ Metrics/ParameterLists:
 # Offense count: 8
 # Configuration parameters: IgnoredMethods.
 Metrics/PerceivedComplexity:
-  Max: 13
+  Max: 12
 
 # Offense count: 11
 Naming/AccessorMethodName:
@@ -141,7 +141,7 @@ Rails/FilePath:
     - 'app/services/marc_service.rb'
 
 # Offense count: 5
-# This cop supports safe auto-correction (--auto-correct).
+# This cop supports unsafe autocorrection (--autocorrect-all).
 # Configuration parameters: EnforcedStyle.
 # SupportedStyles: strict, flexible
 Rails/TimeZone:
@@ -158,14 +158,6 @@ Style/CombinableLoops:
   Exclude:
     - 'app/services/digital_stacks_service.rb'
 
-# Offense count: 1
-# This cop supports safe auto-correction (--auto-correct).
-# Configuration parameters: EnforcedStyle, SingleLineConditionsOnly, IncludeTernaryExpressions.
-# SupportedStyles: assign_to_condition, assign_inside_condition
-Style/ConditionalAssignment:
-  Exclude:
-    - 'spec/support/foxml_helper.rb'
-
 # Offense count: 9
 # Configuration parameters: AllowedConstants.
 Style/Documentation:
@@ -179,13 +171,13 @@ Style/Documentation:
     - 'config/initializers/okcomputer.rb'
 
 # Offense count: 11
-# This cop supports safe auto-correction (--auto-correct).
+# This cop supports safe autocorrection (--autocorrect).
 # Configuration parameters: Strict, AllowedNumbers.
 Style/NumericLiterals:
   MinDigits: 7
 
 # Offense count: 1
-# This cop supports unsafe auto-correction (--auto-correct-all).
+# This cop supports unsafe autocorrection (--autocorrect-all).
 # Configuration parameters: EnforcedStyle, IgnoredMethods.
 # SupportedStyles: predicate, comparison
 Style/NumericPredicate:
@@ -200,7 +192,7 @@ Style/OptionalBooleanParameter:
     - 'app/services/release_tags/identity_metadata.rb'
 
 # Offense count: 7
-# This cop supports safe auto-correction (--auto-correct).
+# This cop supports safe autocorrection (--autocorrect).
 # Configuration parameters: PreferredDelimiters.
 Style/PercentLiteralDelimiters:
   Exclude:
@@ -210,7 +202,7 @@ Style/PercentLiteralDelimiters:
     - 'config/initializers/okcomputer.rb'
 
 # Offense count: 9
-# This cop supports safe auto-correction (--auto-correct).
+# This cop supports safe autocorrection (--autocorrect).
 # Configuration parameters: EnforcedStyle, ConsistentQuotesInMultiline.
 # SupportedStyles: single_quotes, double_quotes
 Style/StringLiterals:
@@ -220,7 +212,7 @@ Style/StringLiterals:
     - 'spec/spec_helper.rb'
 
 # Offense count: 5
-# This cop supports safe auto-correction (--auto-correct).
+# This cop supports safe autocorrection (--autocorrect).
 # Configuration parameters: EnforcedStyle, MinSize.
 # SupportedStyles: percent, brackets
 Style/SymbolArray:
@@ -229,8 +221,8 @@ Style/SymbolArray:
     - 'config/routes.rb'
     - 'lib/tasks/rspec.rake'
 
-# Offense count: 445
-# This cop supports safe auto-correction (--auto-correct).
+# Offense count: 403
+# This cop supports safe autocorrection (--autocorrect).
 # Configuration parameters: AllowHeredoc, AllowURI, URISchemes, IgnoreCopDirectives, AllowedPatterns, IgnoredPatterns.
 # URISchemes: http, https
 Layout/LineLength:
