+++ app/services/cocina_object_store.rb
@@ -30,28 +30,13 @@ class CocinaObjectStore
 
   # Updates a Cocina object in the datastore.
   # @param [Cocina::Models::DRO|Collection|AdminPolicy|DROWithMetadata|CollectionWithMetadata|AdminPolicyWithMetadata] cocina_object
-  # @param [#create] event_factory creates events
   # @param [boolean] skip_lock do not perform an optimistic lock check
   # @raise [Cocina::ValidationError] raised when validation of the Cocina object fails.
   # @raise [CocinaObjectNotFoundError] raised if the cocina object does not already exist in the datastore.
   # @raise [StateLockError] raised if optimistic lock failed.
   # @return [Cocina::Models::AdminPolicyWithMetadata,Cocina::Models::DROWithMetadata,Cocina::Models::CollectionWithMetadata] the saved object with its metadata
-  def self.store(cocina_object, skip_lock:, event_factory: EventFactory)
-    new(event_factory:).store(cocina_object, skip_lock:)
-  end
-
-  # TODO: This method should be renamed to `update'.
-  # Normalizes, validates, and updates a Cocina object in the datastore.
-  # Since normalization is performed, the Cocina object that is returned may differ from the Cocina object that is provided.
-  # @param [Cocina::Models::DRO|Collection|AdminPolicy|DROWithMetadata|CollectionWithMetadata|AdminPolicyWithMetadata] cocina_object
-  # @param [#create] event_factory creates events
-  # @param [boolean] skip_lock do not perform an optimistic lock check
-  # @raise [Cocina::ValidationError] raised when validation of the Cocina object fails.
-  # @raise [CocinaObjectNotFoundError] raised if the cocina object does not already exist in the datastore.
-  # @raise [StateLockError] raised if optimistic lock failed.
-  # @return [Cocina::Models::DRO, Cocina::Models::Collection, Cocina::Models::AdminPolicy] normalized cocina object
-  def self.save(cocina_object, event_factory: EventFactory, skip_lock: false)
-    new(event_factory:).save(cocina_object, skip_lock:)
+  def self.store(cocina_object, skip_lock:)
+    new.store(cocina_object, skip_lock:)
   end
 
   # Removes a Cocina object from the datastore.
@@ -72,10 +57,6 @@ class CocinaObjectStore
     end
   end
 
-  def initialize(event_factory: EventFactory)
-    @event_factory = event_factory
-  end
-
   # @return [Cocina::Models::DROWithMetadata,Cocina::Models::CollectionWithMetadata,Cocina::Models::AdminPolicyWithMetadata]
   def find(druid)
     ar_to_cocina_find(druid)
@@ -83,25 +64,6 @@ class CocinaObjectStore
 
   def store(cocina_object, skip_lock:)
     cocina_to_ar_save(cocina_object, skip_lock:)
-  end
-
-  # TODO: This method should be renamed to "update" since it sends out update messages.
-  def save(cocina_object, skip_lock: false)
-    Cocina::ObjectValidator.validate(cocina_object)
-    # Only update if already exists in PG (i.e., added by create or migration).
-    cocina_object_with_metadata = store(cocina_object, skip_lock:)
-
-    cocina_object_without_metadata = Cocina::Models.without_metadata(cocina_object)
-
-    event_factory.create(druid: cocina_object.externalIdentifier, event_type: 'update', data: { success: true, request: cocina_object_without_metadata.to_h })
-
-    # Broadcast this update action to a topic
-    Notifications::ObjectUpdated.publish(model: cocina_object_with_metadata)
-    cocina_object_with_metadata
-  rescue Cocina::ValidationError => e
-    event_factory.create(druid: cocina_object.externalIdentifier, event_type: 'update',
-                         data: { success: false, error: e.message, request: Cocina::Models.without_metadata(cocina_object).to_h })
-    raise
   end
 
   def exists?(druid)
@@ -122,8 +84,6 @@ class CocinaObjectStore
   end
 
   private
-
-  attr_reader :event_factory
 
   # The *ar* methods are private. In later steps in the migration, the *ar* methods will be invoked by the
   # above public methods.
